+++ geolocator_linux/lib/src/geolocator_linux_method_channel.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/foundation.dart';
+import 'package:flutter/services.dart';
+
+import 'geolocator_linux_native.dart';
+
+class MethodChannelGeolocatorLinux extends GeolocatorLinuxNative {
+  @visibleForTesting
+  final methodChannel = const MethodChannel('geolocator_linux');
+
+  @override
+  Future<bool?> openGnomeLocationSettings() async {
+    return await methodChannel.invokeMethod<bool>('openGnomeLocationPanel');
+  }
+}