+++ geolocator_linux/lib/src/geolocator_linux_native.dart
@@ -1,0 +1,3 @@
+abstract class GeolocatorLinuxNative {
+  Future<bool?> openGnomeLocationSettings();
+}